move status to root mainflux dir
