+++ things/clients/postgres/clients.go
@@ -9,7 +9,9 @@ import (
 	"time"
 
 	"github.com/jackc/pgtype" // required for SQL access
+	mfclients "github.com/mainflux/mainflux/internal/mainflux"
 	mfclients "github.com/mainflux/mainflux/internal/mainflux/clients"
+	"github.com/mainflux/mainflux/internal/mainflux/groups"
 	"github.com/mainflux/mainflux/internal/postgres"
 	"github.com/mainflux/mainflux/pkg/errors"
 	"github.com/mainflux/mainflux/things/clients"
