+++ things/clients/service_test.go
@@ -7,7 +7,8 @@ import (
 	"time"
 
 	"github.com/mainflux/mainflux/internal/apiutil"
-	mfclients "github.com/mainflux/mainflux/internal/mainflux/clients"
+	mfclients "github.com/mainflux/mainflux/internal/mainflux"
+	mferrors "github.com/mainflux/mainflux/internal/mainflux/clients"
 	"github.com/mainflux/mainflux/internal/testsutil"
 	"github.com/mainflux/mainflux/pkg/errors"
 	"github.com/mainflux/mainflux/pkg/uuid"
@@ -816,7 +817,7 @@ func TestEnableClient(t *testing.T) {
 			token:    token,
 			client:   enabledClient1,
 			response: enabledClient1,
-			err:      mfclients.ErrStatusAlreadyAssigned,
+			err:      mferrors.ErrStatusAlreadyAssigned,
 		},
 		{
 			desc:     "enable non-existing client",
@@ -931,7 +932,7 @@ func TestDisableClient(t *testing.T) {
 			token:    token,
 			client:   disabledClient1,
 			response: clients.Client{},
-			err:      mfclients.ErrStatusAlreadyAssigned,
+			err:      mferrors.ErrStatusAlreadyAssigned,
 		},
 		{
 			desc:     "disable non-existing client",
