+++ users/groups/service_test.go
@@ -7,7 +7,8 @@ import (
 	"testing"
 	"time"
 
-	mfclients "github.com/mainflux/mainflux/internal/mainflux/clients"
+	mfclients "github.com/mainflux/mainflux/internal/mainflux"
+	mferrors "github.com/mainflux/mainflux/internal/mainflux/clients"
 	mfgroups "github.com/mainflux/mainflux/internal/mainflux/groups"
 	"github.com/mainflux/mainflux/internal/testsutil"
 	"github.com/mainflux/mainflux/pkg/errors"
@@ -454,7 +455,7 @@ func TestEnableGroup(t *testing.T) {
 			token:    testsutil.GenerateValidToken(t, testsutil.GenerateUUID(t, idProvider), csvc, cRepo, phasher),
 			group:    enabledGroup1,
 			response: enabledGroup1,
-			err:      mfclients.ErrStatusAlreadyAssigned,
+			err:      mferrors.ErrStatusAlreadyAssigned,
 		},
 		{
 			desc:     "enable non-existing group",
@@ -583,7 +584,7 @@ func TestDisableGroup(t *testing.T) {
 			token:    testsutil.GenerateValidToken(t, testsutil.GenerateUUID(t, idProvider), csvc, cRepo, phasher),
 			group:    disabledGroup,
 			response: mfgroups.Group{},
-			err:      mfclients.ErrStatusAlreadyAssigned,
+			err:      mferrors.ErrStatusAlreadyAssigned,
 		},
 		{
 			desc:     "disable non-existing group",
